+++ includes/class-no-img-download-loader.php
@@ -6,8 +6,8 @@
  * @link       http://example.com
  * @since      1.0.0
  *
- * @package    Plugin_Name
- * @subpackage Plugin_Name/includes
+ * @package    No_Img_Download
+ * @subpackage No_Img_Download/includes
  */
 
 /**
@@ -17,11 +17,11 @@
  * the plugin, and register them with the WordPress API. Call the
  * run function to execute the list of actions and filters.
  *
- * @package    Plugin_Name
- * @subpackage Plugin_Name/includes
+ * @package    No_Img_Download
+ * @subpackage No_Img_Download/includes
  * @author     Your Name <email@example.com>
  */
-class Plugin_Name_Loader {
+class No_Img_Download_Loader {
 
 	/**
 	 * The array of actions registered with WordPress.