+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # 1.0.0 
 Initial release
+    - on img tags only
+    - add overlay on load and resize events